+++ webrtc/src/webrtc_c/api/rtp_transceiver_interface.cc
@@ -48,10 +48,11 @@ webrtc_RtpTransceiverInit_get_stream_ids(
 }
 WEBRTC_EXPORT void webrtc_RtpTransceiverInit_set_send_encodings(
     struct webrtc_RtpTransceiverInit* self,
-    struct webrtc_RtpEncodingParameters_vector* encodings) {
+    const struct webrtc_RtpEncodingParameters_vector* encodings) {
   auto init = reinterpret_cast<webrtc::RtpTransceiverInit*>(self);
   auto vec =
-      reinterpret_cast<std::vector<webrtc::RtpEncodingParameters>*>(encodings);
+      reinterpret_cast<const std::vector<webrtc::RtpEncodingParameters>*>(
+          encodings);
   init->send_encodings = *vec;
 }
 }